aleatoire accueil + clean call rouge

+++ conf/evolutions/default/1.sql
@@ -6,7 +6,7 @@
 create table audio (
   id                        bigint auto_increment not null,
   lien                      varchar(255),
-  objet_id                  bigint,
+  objet_id                  integer,
   constraint pk_audio primary key (id))
 ;
 
@@ -19,7 +19,7 @@ create table contenu_site (
 create table favori (
   id                        bigint auto_increment not null,
   utilisateur_id            bigint,
-  objet_id                  bigint,
+  objet_id                  integer,
   constraint pk_favori primary key (id))
 ;
 
@@ -34,19 +34,19 @@ create table feedback (
 create table historique (
   id                        bigint auto_increment not null,
   utilisateur_id            bigint,
-  objet_id                  bigint,
+  objet_id                  integer,
   constraint pk_historique primary key (id))
 ;
 
 create table image (
   id                        bigint auto_increment not null,
   lien                      varchar(255),
-  objet_id                  bigint,
+  objet_id                  integer,
   constraint pk_image primary key (id))
 ;
 
 create table objet (
-  id                        bigint auto_increment not null,
+  id                        integer auto_increment not null,
   nom                       varchar(255),
   reference                 varchar(255),
   description               TEXT,
@@ -74,7 +74,7 @@ create table parcours (
 
 create table parcours_objet (
   id                        bigint auto_increment not null,
-  objet_id                  bigint,
+  objet_id                  integer,
   parcours_id               bigint,
   constraint pk_parcours_objet primary key (id))
 ;
@@ -92,20 +92,20 @@ create table utilisateur (
 create table video (
   id                        bigint auto_increment not null,
   lien                      varchar(255),
-  objet_id                  bigint,
+  objet_id                  integer,
   constraint pk_video primary key (id))
 ;
 
 
 create table statistique_objet (
-  objet1                         bigint not null,
-  objet2                         bigint not null,
+  objet1                         integer not null,
+  objet2                         integer not null,
   constraint pk_statistique_objet primary key (objet1, objet2))
 ;
 
 create table oeuvre_composite (
-  oeuvre_principale              bigint not null,
-  oeuvre_inspiree                bigint not null,
+  oeuvre_principale              integer not null,
+  oeuvre_inspiree                integer not null,
   constraint pk_oeuvre_composite primary key (oeuvre_principale, oeuvre_inspiree))
 ;
 alter table audio add constraint fk_audio_objet_1 foreign key (objet_id) references objet (id) on delete restrict on update restrict;
